Large refactoring to flatten structures hierarchy
