--- src/misc/mod.rs
@@ -1,9 +1,0 @@
-//! Structures not included in other groups
-//!
-
-pub mod oem_strings;
-pub use self::oem_strings::OemStrings;
-
-pub mod group_associations;
-pub use self::group_associations::GroupAssociations;
-
